removed old fetchActiviyLocations
